// generated/src/app/api-models/manual-check-select-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ManualCheckSelectWork {
  seqCkprtId: number;
  jobId: string;
  passNumber: number;
  checkNumber: string;
  seqApTrans: number;
  requestUser: string;
  runDate: Date;
  payType: string;
  companyCode: string;
  status: string;
  seqVendorId: number;
  seqVendorAddress: number;
  seqProviderId: number;
  seqMemberId: number;
  seqClaimId: number;
  detailSvcDate: Date;
  procedureCode: string;
  netAmt: number;
  capMonth: Date;
  tranType: string;
  tranAmt: number;
  capModelId: string;
  capEntityCode: string;
  seqCapPoolId: number;
  fromPostDate: Date;
  thruPostDate: Date;
  fromDueDate: Date;
  thruDueDate: Date;
  includeFlag: string;
  seqCapTrans: number;
  lineNumber: number;
  subLineCode: string;
}